break cube into own class
