+++ lib/index.js
@@ -1,47 +1,7 @@
 var canvas = document.getElementById("canvas");
 var context = canvas.getContext('2d');
+var Cube = require('./cube');
 
-// CUBE CONSTRUCTOR //
-function Cube(params) {
-  this.active       = false;
-  this.id           = params["id"];
-  this.x            = params["x"];
-  this.y            = params["y"];
-  this.upLeftId     = params["upLeftId"];
-  this.upRightId    = params["upRightId"];
-  this.downLeftId   = params["downLeftId"];
-  this.downRightId  = params["downRightId"];
-}
-
-Cube.prototype.drawCube = function(status){
-  if (status === 'active'){
-    var topColor = '#FFCC46';
-  } else {
-    var topColor = '#3B6D80';
-  };
-
-  context.beginPath();
-  context.moveTo(this.x, this.y);
-  context.lineTo(this.x+40, this.y+10);
-  context.lineTo(this.x, this.y+20);
-  context.lineTo(this.x-40, this.y+10);
-  context.fillStyle= topColor;
-  context.fill();
-  context.beginPath();
-  context.moveTo(this.x, this.y+20);
-  context.lineTo(this.x, this.y+70);
-  context.lineTo(this.x+40, this.y+60);
-  context.lineTo(this.x+40, this.y+10);
-  context.fillStyle="#47528B";
-  context.fill();
-  context.beginPath();
-  context.moveTo(this.x, this.y+20);
-  context.lineTo(this.x, this.y+70);
-  context.lineTo(this.x-40, this.y+60);
-  context.lineTo(this.x-40, this.y+10);
-  context.fillStyle="#529A68";
-  context.fill();
-};
 
 // CREATE CUBES //
 var cubes = [];
@@ -56,6 +16,7 @@ for(var i = 0; i < 7; i++){
       upRightId: j < i ? cubes.length - i : null,
       downLeftId: cubes.length + (i + 1) <= 27 ? cubes.length + (i + 1) : null,
       downRightId: cubes.length + (i + 2) <= 27 ? cubes.length + (i + 2) : null,
+      context: context
     };
     var cube = new Cube(newParams);
     cubes.push(cube);
